Moved notification plugin to separate directory.
